expose facingMode functions
