+++ src/index.ts
@@ -37,6 +37,7 @@ export * from './room/events';
 export * from './room/track/Track';
 export * from './room/track/create';
 export * from './room/track/options';
+export { facingModeFromDeviceLabel, facingModeFromLocalTrack } from './room/track/facingMode';
 export * from './room/track/types';
 export type { DataPublishOptions, SimulationScenario } from './room/types';
 export * from './version';
