+++ src/room/track/utils.test.ts
@@ -1,5 +1,5 @@
 import { AudioCaptureOptions, VideoCaptureOptions, VideoPresets } from './options';
-import { constraintsForOptions, facingModeFromDeviceLabel, mergeDefaultOptions } from './utils';
+import { constraintsForOptions, mergeDefaultOptions } from './utils';
 
 describe('mergeDefaultOptions', () => {
   const audioDefaults: AudioCaptureOptions = {
@@ -106,34 +106,5 @@ describe('constraintsForOptions', () => {
     expect(videoOpts.height).toEqual(VideoPresets.h720.resolution.height);
     expect(videoOpts.frameRate).toEqual(VideoPresets.h720.resolution.frameRate);
     expect(videoOpts.aspectRatio).toEqual(VideoPresets.h720.resolution.aspectRatio);
-  });
-});
-
-describe('Test facingMode detection', () => {
-  test('OBS virtual camera should be detected.', () => {
-    const result = facingModeFromDeviceLabel('OBS Virtual Camera');
-    expect(result?.facingMode).toEqual('environment');
-    expect(result?.confidence).toEqual('medium');
-  });
-
-  test.each([
-    ['Peter’s iPhone Camera', { facingMode: 'environment', confidence: 'medium' }],
-    ['iPhone de Théo Camera', { facingMode: 'environment', confidence: 'medium' }],
-  ])(
-    'Device labels that contain "iphone" should return facingMode "environment".',
-    (label, expected) => {
-      const result = facingModeFromDeviceLabel(label);
-      expect(result?.facingMode).toEqual(expected.facingMode);
-      expect(result?.confidence).toEqual(expected.confidence);
-    },
-  );
-
-  test.each([
-    ['Peter’s iPad Camera', { facingMode: 'environment', confidence: 'medium' }],
-    ['iPad de Théo Camera', { facingMode: 'environment', confidence: 'medium' }],
-  ])('Device label that contain "ipad" should detect.', (label, expected) => {
-    const result = facingModeFromDeviceLabel(label);
-    expect(result?.facingMode).toEqual(expected.facingMode);
-    expect(result?.confidence).toEqual(expected.confidence);
   });
 });
